UPSTREAM distribution/distribution: 3936: registry/storage/driver/azure: fix Move method
Something seems broken on azure/azure sdk side - it is currently not
possible to copy a blob of type AppendBlob using `CopyFromURL`.
Using the AppendBlob client via NewAppendBlobClient does not work
either.
The only thing that did work was using the async copy via
StartCopyFromURL, so that is what is used here.

Signed-off-by: Flavian Missi <fmissi@redhat.com>